# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tabi API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/checkins/{checkin_id}": {
      "get": {
        "tags": [
          "checkins"
        ],
        "summary": "Own Detail",
        "description": "Read one owned experience.",
        "operationId": "own_detail_api_checkins__checkin_id__get",
        "parameters": [
          {
            "name": "checkin_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Checkin Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnCheckinOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "checkins"
        ],
        "summary": "Edit Checkin",
        "description": "Edit one owned experience without affecting other visits.",
        "operationId": "edit_checkin_api_checkins__checkin_id__patch",
        "parameters": [
          {
            "name": "checkin_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Checkin Id"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckinEdit"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnCheckinOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "checkins"
        ],
        "summary": "Delete Checkin",
        "description": "Soft-delete one record and immediately remove it from progress/shares.",
        "operationId": "delete_checkin_api_checkins__checkin_id__delete",
        "parameters": [
          {
            "name": "checkin_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Checkin Id"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/media/uploads": {
      "post": {
        "tags": [
          "media"
        ],
        "summary": "Upload Photo",
        "description": "Stage one validated image for a future experience.",
        "operationId": "upload_photo_api_media_uploads_post",
        "parameters": [
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_photo_api_media_uploads_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Upload Photo Api Media Uploads Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/media/checkins/{checkin_id}": {
      "post": {
        "tags": [
          "media"
        ],
        "summary": "Add Photo",
        "description": "Append a validated image to one owned experience.",
        "operationId": "add_photo_api_media_checkins__checkin_id__post",
        "parameters": [
          {
            "name": "checkin_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Checkin Id"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_add_photo_api_media_checkins__checkin_id__post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Add Photo Api Media Checkins  Checkin Id  Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/media/{media_id}": {
      "delete": {
        "tags": [
          "media"
        ],
        "summary": "Remove Photo",
        "description": "Remove one owned photo while preserving nonempty experiences.",
        "operationId": "remove_photo_api_media__media_id__delete",
        "parameters": [
          {
            "name": "media_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Media Id"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/media/{media_id}/thumbnail": {
      "get": {
        "tags": [
          "media"
        ],
        "summary": "Thumbnail",
        "description": "Serve a thumbnail to its owner or while the experience is public.",
        "operationId": "thumbnail_api_media__media_id__thumbnail_get",
        "parameters": [
          {
            "name": "media_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Media Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/media/{media_id}/original": {
      "get": {
        "tags": [
          "media"
        ],
        "summary": "Original",
        "description": "Serve a processed full-size image only to its author.",
        "operationId": "original_api_media__media_id__original_get",
        "parameters": [
          {
            "name": "media_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Media Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/media/lists/{list_id}/cover": {
      "post": {
        "tags": [
          "media"
        ],
        "summary": "List Cover",
        "description": "Set a checklist's processed cover image.",
        "operationId": "list_cover_api_media_lists__list_id__cover_post",
        "parameters": [
          {
            "name": "list_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "List Id"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_list_cover_api_media_lists__list_id__cover_post"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "media"
        ],
        "summary": "Get List Cover",
        "description": "Serve a published checklist cover.",
        "operationId": "get_list_cover_api_media_lists__list_id__cover_get",
        "parameters": [
          {
            "name": "list_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "List Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/media/items/{item_id}/cover": {
      "post": {
        "tags": [
          "media"
        ],
        "summary": "Item Cover",
        "description": "Set an item's processed cover image.",
        "operationId": "item_cover_api_media_items__item_id__cover_post",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_item_cover_api_media_items__item_id__cover_post"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "media"
        ],
        "summary": "Get Item Cover",
        "description": "Serve a published item's cover.",
        "operationId": "get_item_cover_api_media_items__item_id__cover_get",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/media/avatar": {
      "post": {
        "tags": [
          "media"
        ],
        "summary": "Avatar",
        "description": "Set the signed-in account's public avatar.",
        "operationId": "avatar_api_media_avatar_post",
        "parameters": [
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_avatar_api_media_avatar_post"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/media/avatar/{user_id}": {
      "get": {
        "tags": [
          "media"
        ],
        "summary": "Get Avatar",
        "description": "Serve a processed public avatar without exposing account details.",
        "operationId": "get_avatar_api_media_avatar__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_add_photo_api_media_checkins__checkin_id__post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_add_photo_api_media_checkins__checkin_id__post"
      },
      "Body_avatar_api_media_avatar_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_avatar_api_media_avatar_post"
      },
      "Body_item_cover_api_media_items__item_id__cover_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_item_cover_api_media_items__item_id__cover_post"
      },
      "Body_list_cover_api_media_lists__list_id__cover_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_list_cover_api_media_lists__list_id__cover_post"
      },
      "Body_upload_photo_api_media_uploads_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_photo_api_media_uploads_post"
      },
...
      "CheckinEdit": {
        "properties": {
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 30000
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "experienced_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experienced At"
          },
          "visibility": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "private",
                  "public"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Visibility"
          },
          "position": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PositionIn"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "CheckinEdit",
        "description": "Fields editable on one existing experience."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MediaOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "thumbnail_url": {
            "type": "string",
            "title": "Thumbnail Url"
          },
          "original_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Url"
          },
          "sort_order": {
            "type": "integer",
            "title": "Sort Order"
          }
        },
        "type": "object",
        "required": [
          "id",
          "thumbnail_url",
          "original_url",
          "sort_order"
        ],
        "title": "MediaOut",
        "description": "Safe media URLs bound to one experience."
      },
      "OwnCheckinOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "share_id": {
            "type": "string",
            "title": "Share Id"
          },
          "item_id": {
            "type": "string",
            "title": "Item Id"
          },
          "item_name": {
            "type": "string",
            "title": "Item Name"
          },
          "list_id": {
            "type": "string",
            "title": "List Id"
          },
          "list_title": {
            "type": "string",
            "title": "List Title"
          },
          "author_id": {
            "type": "string",
            "title": "Author Id"
          },
          "author_name": {
            "type": "string",
            "title": "Author Name"
          },
          "author_avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author Avatar Url"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "experienced_at": {
            "type": "string",
            "format": "date-time",
            "title": "Experienced At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "media": {
            "items": {
              "$ref": "#/components/schemas/MediaOut"
            },
            "type": "array",
            "title": "Media"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "private",
              "public"
            ],
            "title": "Visibility"
          },
          "hidden": {
            "type": "boolean",
            "title": "Hidden"
          },
          "position": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PositionIn"
              },
              {
                "type": "null"
              }
            ]
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "item_status": {
            "type": "string",
            "title": "Item Status"
          },
          "list_status": {
            "type": "string",
            "title": "List Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "share_id",
          "item_id",
          "item_name",
          "list_id",
          "list_title",
          "author_id",
          "author_name",
          "author_avatar_url",
          "note",
          "experienced_at",
          "created_at",
          "media",
          "visibility",
          "hidden",
          "position",
          "updated_at",
          "item_status",
          "list_status"
        ],
        "title": "OwnCheckinOut",
        "description": "Owner view including private state and optional position."
      },
...
    }
  }
}